refactor: eliminate extract_benchmark_name duplication
Move extract_benchmark_name function to sightglass-data crate as shared
utility. Update CLI and analysis crates to use the shared function,
removing duplicate implementations and tests.
